Finish inheritance; Figured out that inheritance is not working well here because of template arguments and different return variables in model class; Implemented one-size-fit-all run.cpp

+++ include/utilities/parameter_reader.hpp
@@ -19,10 +19,10 @@
 namespace params
 {
 	extern std::string file_name;
-	extern int horizon;
+	extern size_t horizon;
 	extern double epsilon;
-	extern int start_state;
-	extern std::vector<int> start_belief;
+	extern size_t start_state;
+	extern std::vector<size_t> start_belief;
 	extern std::string solver_type;
 	extern std::string model_type;
 	extern std::string model_scale;
